refactor: 시딩 인프라를 global/seed 에서 test 도메인으로 이전
ApplicationRunner 기반 부팅 시 자동 시딩(AlphaSeedRunner)을 폐기하고,
DummyMemberFactory 와 SeedProperties 를 test/application/service 로 이전한다.
Alpha 접두사를 제거해 alpha 한정 정체성을 지우고, properties 키 app.seed.alpha.*
도 app.seed.* 로 일반화한다. 가드는 @Profile("!prod") + app.seed.enabled 로
prod 제외 전 환경에서 활성화한다.

기존 LoadMemberPort 직접 의존(AlphaSeedRunner.run 의 countAllMembers 호출)은
Runner 폐기로 자연 소멸한다. ADR-017 참조.
