+++ src/main/java/com/umc/product/test/application/service/SeedProperties.java
@@ -1,18 +1,18 @@
-package com.umc.product.global.seed;
+package com.umc.product.test.application.service;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
 /**
- * alpha 환경 더미 데이터 시딩 설정. ADR-007 참조.
+ * test 도메인 시딩 API 설정. ADR-017 참조.
  * <p>
  * 활성화는 두 조건을 모두 만족해야 한다.
  * <ol>
- *   <li>spring.profiles.active 가 alpha 를 포함</li>
- *   <li>app.seed.alpha.enabled = true</li>
+ *   <li>spring.profiles.active 가 prod 가 아닐 것</li>
+ *   <li>app.seed.enabled = true</li>
  * </ol>
  */
-@ConfigurationProperties(prefix = "app.seed.alpha")
-public record AlphaSeedProperties(
+@ConfigurationProperties(prefix = "app.seed")
+public record SeedProperties(
     boolean enabled,
     long skipIfMemberCountGreaterThan,
     int idPwMemberCount,
