+++ src/main/resources/application.yml
@@ -338,16 +338,15 @@ app:
     username: ${SWAGGER_AUTH_USERNAME:username}
     password: ${SWAGGER_AUTH_PASSWORD:password}
 
-  # alpha 환경 전용 더미 데이터 시딩. ADR-007 참조.
-  # 기본값은 비활성. profile=alpha 일 때 하단 multi-document 섹션이 enabled=true 로 오버라이드한다.
+  # test 도메인 시딩 API 설정. ADR-017 참조.
+  # 기본값은 비활성. prod 가 아닌 환경에서 APP_SEED_ENABLED=true 로 켜야 시딩 API 가 노출된다.
   seed:
-    alpha:
-      enabled: ${APP_SEED_ALPHA_ENABLED:false}
-      skip-if-member-count-greater-than: ${APP_SEED_ALPHA_SKIP_THRESHOLD:0}
-      id-pw-member-count: ${APP_SEED_ALPHA_IDPW_COUNT:20}
-      oauth-member-count: ${APP_SEED_ALPHA_OAUTH_COUNT:30}
-      email-domain: ${APP_SEED_ALPHA_EMAIL_DOMAIN:alpha.umc.test}
-      default-password: ${APP_SEED_ALPHA_DEFAULT_PASSWORD:Alpha!Pass2026}
+    enabled: ${APP_SEED_ENABLED:false}
+    skip-if-member-count-greater-than: ${APP_SEED_SKIP_THRESHOLD:0}
+    id-pw-member-count: ${APP_SEED_IDPW_COUNT:20}
+    oauth-member-count: ${APP_SEED_OAUTH_COUNT:30}
+    email-domain: ${APP_SEED_EMAIL_DOMAIN:alpha.umc.test}
+    default-password: ${APP_SEED_DEFAULT_PASSWORD:Alpha!Pass2026}
 
   oauth2:
     google:
@@ -452,6 +451,5 @@ spring:
 
 app:
   seed:
-    alpha:
-      # alpha 에서는 기본 활성. 일시 비활성하고 싶다면 APP_SEED_ALPHA_ENABLED=false 로 설정.
-      enabled: ${APP_SEED_ALPHA_ENABLED:true}
+    # alpha 에서는 기본 활성. 일시 비활성하고 싶다면 APP_SEED_ENABLED=false 로 설정.
+    enabled: ${APP_SEED_ENABLED:true}
